pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "intake_oceantide"
description = "Intake driver for oceantide"
readme = "README.rst"
authors = [
  {name = "Oceanum Developers", email = "developers@oceanum.science"}
]
maintainers = [
  {name = "Oceanum Developers", email = "developers@oceanum.science"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "fsspec",
  "intake-xarray",
  "numpy",
  "oceantide",
  "pandas",
  "typer",
  "xarray",
  "zarr",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff",  # linting
    "sphinx",  # documentation
    "tox",  # testing
    "twine",  # publishing
]

[project.urls]

bugs = "https://github.com/oceanum/intake_oceantide/issues"
changelog = "https://github.com/oceanum/intake_oceantide/blob/master/changelog.md"
homepage = "https://github.com/oceanum/intake_oceantide"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

